+++ dhondtxai/__init__.py
@@ -1,12 +1,15 @@
 
-from .dhondt_xai import DhondtExplanation, DhondtXAI
+from .dhondt_xai import DhondtExplanation, DhondtValues, DhondtXAI
 from .plot_parliament import plot_parliament, plot_signed_parliament
 
-__version__ = "0.9.1"
+__version__ = "0.9.2"
+Explainer = DhondtXAI
 
 __all__ = [
     "DhondtExplanation",
+    "DhondtValues",
     "DhondtXAI",
+    "Explainer",
     "plot_parliament",
     "plot_signed_parliament",
 ]